code lint added

+++ contracts/Constructor-16.sol
@@ -4,9 +4,9 @@ pragma solidity ^0.8.7;
 
 contract Constructor {
     address public owner;
-    uint public x;
+    uint256 public x;
 
-    constructor(uint _x) {
+    constructor(uint256 _x) {
         owner = msg.sender;
         x = _x;
     }
